# Feature/20161012 speed optimization