--- a/tools/performance_cehck.zsh
+++ b/tools/performance_cehck.zsh
@@ -1,15 +1,30 @@
 #! /usr/bin/env zsh
 
-function nano_unix_time() {
-    date +%s%N
+function milliseconds_unix_time() {
+    echo $(($(date +%s%N) / 1000000.0))
 }
 
 function show_proc_time() {
     cmd=${1}
 
-    start_t=$(nano_unix_time)
-    ${cmd} 1>/dev/null 2>/dev/null
-    printf "%40s : time -> %10d\n" ${cmd} $(($(nano_unix_time) - ${start_t}))
+    local avg_time
+    avg_time=0
+
+    local number_of_samples=5
+
+    local start_time elapsed_time
+    for i ({1..${number_of_samples}})
+    do
+      start_time=$(milliseconds_unix_time)
+      ${cmd} 1>/dev/null 2>/dev/null
+      elapsed_time=$(($(milliseconds_unix_time) - ${start_time}))
+
+      avg_time=$((${avg_time} + ${elapsed_time}))
+    done
+
+    avg_time=$((${avg_time} / ${number_of_samples}))
+
+    printf "%40s : time (%d times avg) -> %5f ms\n" ${cmd:t} ${number_of_samples} ${avg_time}
 }
 
 # Load