This PR solves performance issue.

__Speed comparison__

_Original_

```
                         git-added-count : time (5 times avg) -> 7.360303 ms
                     git-untracked-count : time (5 times avg) -> 6.853711 ms
                       git-in-repository : time (5 times avg) -> 3.549219 ms
          git-commit-diff-current-master : time (5 times avg) -> 8.035645 ms
                 git-current-branch-name : time (5 times avg) -> 5.400195 ms
    git-commit-diff-current-master-count : time (5 times avg) -> 11.993994 ms
                   git-update-count-base : time (5 times avg) -> 2.118311 ms
                         git-commit-diff : time (5 times avg) -> 1.636670 ms
                      git-modified-count : time (5 times avg) -> 6.434766 ms
                       git-deleted-count : time (5 times avg) -> 5.930029 ms
                      git-colorized-info : time (5 times avg) -> 26.713623 ms
                         git-stash-count : time (5 times avg) -> 15.679541 ms
```

_Modified_

```
                         git-added-count : time (5 times avg) -> 7.318555 ms
                     git-untracked-count : time (5 times avg) -> 6.583447 ms
                       git-in-repository : time (5 times avg) -> 2.878174 ms
          git-commit-diff-current-master : time (5 times avg) -> 7.960889 ms
                 git-current-branch-name : time (5 times avg) -> 4.419824 ms
    git-commit-diff-current-master-count : time (5 times avg) -> 10.091016 ms
                   git-update-count-base : time (5 times avg) -> 2.054980 ms
                         git-commit-diff : time (5 times avg) -> 1.469238 ms
                      git-modified-count : time (5 times avg) -> 6.378125 ms
                       git-deleted-count : time (5 times avg) -> 5.581348 ms
                      git-colorized-info : time (5 times avg) -> 15.319336 ms
                         git-stash-count : time (5 times avg) -> 4.302832 ms
```